refactor: rename PlaybackService > RuntimeService
